+++ sdk/core/azure-core/CHANGELOG.md
@@ -10,7 +10,7 @@
 
 ### Other Changes
 
-- Use default ssl context to be compatible with new aiohttp.
+- Use ssl context from aiohttp by default.
 
 ## 1.29.4 (2023-09-07)